Multiple descriptor sets on the same material
Multiple descriptor sets on the same material should now be possible...
But I didn't check. Anyways, descriptor sets are now a part of material
render objects.

Now that I think of it, There is a problem with descriptor pools. RN
they assume only one descriptor set for each material layout. Which is a
problem. Need to fix it. Too lazy to do this ATM